Create Priyanka and Toys
Priyanka works for an international toy company that ships by container. Her task is to the determine the lowest cost way to combine her orders for shipping. She has a list of item weights. The shipping company has a requirement that all items loaded in a container must weigh less than or equal to 4 units plus the weight of the minimum weight item. All items meeting that requirement will be shipped in one container.

What is the smallest number of containers that can be contracted to ship the items based on the given list of weights?

Sample Output

cost=[2, 2,4, 3]

14

12

Explanation

Sunny and Johnny make the following two trips to the parlor:

1. The first time, they pool together m = 4 dollars. Of the five flavors available that day, flavors 1 and 4

have a total cost of 1+3 = 4.

2. The second time, they pool together m = 4 dollars. Of the four flavors available that day, flavors 1 and 2 have a total cost of 2+2=4.