Create Max Min
You will be given a list of integers, arr, and a single integer k. You must create an array of length k from elements of arr such that its unfairness is minimized. Call that array subarr. Unfairness of an array is calculated as:

unfairness = max(subarr) - min(subarr)

Where:

max(subarr) is the maximum element of subarr,
min(subarr) is the minimum element of subarr.
Function Description
Complete the function maxMin in the editor below.

maxMin has the following parameters:

int k: the number of elements to select
int arr[n]: an array of integers
Returns
int: the minimum possible unfairness
Input Format
The first line contains an integer n, the number of elements in the array arr. The second line contains an integer k. Each of the next n lines contains an integer arr[i] where 0 <= i < n.

Constraints
2 ≤ n ≤ 10^5
2 ≤ k ≤ n
0 ≤ arr[i] ≤ 10^9

Sample Input

7
3
10
100
300
200
1000
20
30

Sample Output

20

Explanation
Here k = 3, and selecting the subarray [10, 20, 30] gives the smallest difference between the maximum and minimum values, which is 30 - 10 = 20.