Create Winning Lottery Ticket
The problem is to determine how many pairs of tickets can be combined to form a winning lottery ticket. A winning lottery ticket is defined as a ticket pair such that the union of the digits on both tickets contains all digits from 0 to 9 at least once.

Function Description
Complete the function winningLotteryTicket in the editor below.

winningLotteryTicket has the following parameter(s):

string tickets[n]: an array of ticket numbers represented as strings.

Returns
long: the number of winning pairs of tickets.

Input Format
The first line contains an integer n, the number of tickets.

Each of the next n lines contains a string tickets[i], a ticket number consisting of digits.

Constraints
1 ≤ n ≤ 10^6
Each ticket consists of digits only and has a length of at most 10 characters.

Sample Input
4
129300455
5559948277
012334556
56789

Sample Output
5

Explanation
There are 5 pairs of tickets where the union of the digits in both tickets contains all digits from 0 to 9:

129300455 + 5559948277
129300455 + 012334556
129300455 + 56789
5559948277 + 012334556
5559948277 + 56789
These pairs of tickets together have all digits from 0 to 9, and hence they are winning pairs.